+++ index.js
@@ -1,8 +1,8 @@
-import { default as Sidebar, moduleNow, tabs } from './src/index.vue';
-
-export { Sidebar, moduleNow, tabs };
-
-
 import TabAdmin, { Tab } from './src/tab-admin.js';
 
 export { TabAdmin, Tab };
+
+
+import { default as Sidebar, moduleNow, tabAdmin, domSidebar } from './src/index.vue';
+
+export { Sidebar, moduleNow, tabAdmin, domSidebar };